Revert "[core] Deprecate LIFO/FIFO worker killing policies" (#56960)
Reverts ray-project/ray#56314

+++ src/ray/common/task/task_spec.cc
@@ -187,7 +187,7 @@ bool TaskSpecification::HasRuntimeEnv() const {
   return !IsRuntimeEnvEmpty(SerializedRuntimeEnv());
 }
 
-int32_t TaskSpecification::AttemptNumber() const { return message_->attempt_number(); }
+uint64_t TaskSpecification::AttemptNumber() const { return message_->attempt_number(); }
 
 bool TaskSpecification::IsRetry() const { return AttemptNumber() > 0; }
